Segregates classes into modules of their domain
