+++ django_date_extensions/fields.py
@@ -1,88 +1,14 @@
 from __future__ import absolute_import
 
 import datetime
-import time
 import re
-from functools import total_ordering
 
 from django.db import models
-from django import forms
 from django.forms import ValidationError
 from django.utils import dateformat
 
-from django_date_extensions import settings
-from django_date_extensions.widgets import PrettyDateInput
-
-
-@total_ordering
-class ApproximateDate:
-    """A date object that accepts 0 for month or day to mean we don't
-       know when it is within that month/year."""
-    def __init__(self, year=0, month=0, day=0, future=False, past=False):
-        if future and past:
-            raise ValueError("Can't be both future and past")
-        elif future or past:
-            if year or month or day:
-                raise ValueError("Future or past dates can have no year, month or day")
-        elif year and month and day:
-            datetime.date(year, month, day)
-        elif year and month:
-            datetime.date(year, month, 1)
-        elif year and day:
-            raise ValueError("You cannot specify just a year and a day")
-        elif year:
-            datetime.date(year, 1, 1)
-        else:
-            raise ValueError("You must specify a year")
-
-        self.future = future
-        self.past = past
-        self.year = year
-        self.month = month
-        self.day = day
-
-    def __repr__(self):
-        if self.future or self.past:
-            return str(self)
-        return "{year:04d}-{month:02d}-{day:02d}".format(year=self.year, month=self.month, day=self.day)
-
-    def __str__(self):
-        if self.future:
-            return 'future'
-        if self.past:
-            return 'past'
-        elif self.year and self.month and self.day:
-            return dateformat.format(self, settings.OUTPUT_FORMAT_DAY_MONTH_YEAR)
-        elif self.year and self.month:
-            return dateformat.format(self, settings.OUTPUT_FORMAT_MONTH_YEAR)
-        elif self.year:
-            return dateformat.format(self, settings.OUTPUT_FORMAT_YEAR)
-
-    def __eq__(self, other):
-        if isinstance(other, (datetime.date, datetime.datetime)):
-            return (self.year, self.month, self.day) ==\
-                   (other.year, other.month, other.day)
-
-        if not isinstance(other, ApproximateDate):
-            return False
-
-        return (self.year, self.month, self.day, self.future, self.past) ==\
-               (other.year, other.month, other.day, other.future, other.past)
-
-    def __lt__(self, other):
-        if other is None:
-            return False
-
-        if isinstance(other, ApproximateDate):
-            if self.future or other.future:
-                return not self.future
-            if self.past or other.past:
-                return not other.past
-
-        return (self.year, self.month, self.day) < (other.year, other.month, other.day)
-
-    def __len__(self):
-        return len(self.__repr__())
+from django_date_extensions.forms import ApproximateDateFormField
+from django_date_extensions.types import ApproximateDate
 
 
 ansi_date_re = re.compile(r'^\d{4}-\d{1,2}-\d{1,2}$')
@@ -153,91 +79,4 @@ class ApproximateDateField(models.CharField):
         return super(ApproximateDateField, self).formfield(**kwargs)
 
 
-# TODO: Expand to work more like my PHP strtotime()-using function
-class ApproximateDateFormField(forms.fields.Field):
-    def __init__(self, max_length=10, empty_value='', *args, **kwargs):
-        super(ApproximateDateFormField, self).__init__(*args, **kwargs)
-
-    def clean(self, value):
-        super(ApproximateDateFormField, self).clean(value)
-        if value in (None, ''):
-            return None
-        if value == 'future':
-            return ApproximateDate(future=True)
-        if value == 'past':
-            return ApproximateDate(past=True)
-        if isinstance(value, ApproximateDate):
-            return value
-        value = re.sub(r'(?<=\d)(st|nd|rd|th)', '', value.strip())
-        for date_format in settings.DATE_INPUT_FORMATS:
-            try:
-                return ApproximateDate(*time.strptime(value, date_format)[:3])
-            except ValueError:
-                continue
-        for month_format in settings.MONTH_INPUT_FORMATS:
-            try:
-                match = time.strptime(value, month_format)
-                return ApproximateDate(match[0], match[1], 0)
-            except ValueError:
-                continue
-        for year_format in settings.YEAR_INPUT_FORMATS:
-            try:
-                return ApproximateDate(time.strptime(value, year_format)[0], 0, 0)
-            except ValueError:
-                continue
-        raise ValidationError('Please enter a valid date.')
-
-
-# PrettyDateField - same as DateField but accepts slightly more input,
-# like ApproximateDateFormField above. If initialised with future=True,
-# it will assume a date without year means the current year (or the next
-# year if the day is before the current date). If future=False, it does
-# the same but in the past.
-class PrettyDateField(forms.fields.Field):
-    widget = PrettyDateInput
-
-    def __init__(self, future=None, *args, **kwargs):
-        self.future = future
-        super(PrettyDateField, self).__init__(*args, **kwargs)
-
-    def clean(self, value):
-        """
-        Validates that the input can be converted to a date. Returns a Python
-        datetime.date object.
-        """
-        super(PrettyDateField, self).clean(value)
-        if value in (None, ''):
-            return None
-        if value == 'future':
-            return ApproximateDate(future=True)
-        if value == 'past':
-            return ApproximateDate(past=True)
-        if isinstance(value, datetime.datetime):
-            return value.date()
-        if isinstance(value, datetime.date):
-            return value
-        value = re.sub(r'(?<=\d)(st|nd|rd|th)', '', value.strip())
-        for date_input_format in settings.DATE_INPUT_FORMATS:
-            try:
-                return datetime.date(*time.strptime(value, date_input_format)[:3])
-            except ValueError:
-                continue
-
-        if self.future is None:
-            raise ValidationError('Please enter a valid date.')
-
-        # Allow year to be omitted. Do the sensible thing, either past or future.
-        for day_month_input_format in settings.DAY_MONTH_INPUT_FORMATS:
-            try:
-                t = time.strptime(value, day_month_input_format)
-                month, day, yday = t[1], t[2], t[7]
-                year = datetime.date.today().year
-                if self.future and yday < int(datetime.date.today().strftime('%j')):
-                    year += 1
-                if not self.future and yday > int(datetime.date.today().strftime('%j')):
-                    year -= 1
-                return datetime.date(year, month, day)
-            except ValueError:
-                continue
-
-        raise ValidationError('Please enter a valid date.')
+__all__ = (ApproximateDateField.__name__)
